+++ src/stories/Card.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react';
-import { Button, Card, CardContent, CardFooter, CardTitle, ConsoleButton } from '../components';
+import { Card, CardContent, CardFooter, CardTitle } from '../components';
 import '../style/globals.css';
 
 // More on how to set up stories at: https://storybook.js.org/docs/writing-stories#default-export
@@ -10,10 +10,8 @@ const meta: Meta<typeof Card> = {
   // More on argTypes: https://storybook.js.org/docs/api/argtypes
   argTypes: {
     variant: {
-      control: {
-        type: 'select',
-        options: ['default', 'clickable', 'console'],
-      },
+      control: 'select',
+      options: ['default', 'clickable', 'console'],
     },
     asChild: {
       control: {
@@ -21,24 +19,12 @@ const meta: Meta<typeof Card> = {
       },
     },
     children: {
-      control: {
-        accept: 'object',
-      },
-    }
+      control: 'object',
+    },
   },
   args: {
     variant: 'default',
     asChild: false,
-  }
-};
-
-export default meta;
-type Story = StoryObj<typeof meta>;
-
-// More on writing stories with args: https://storybook.js.org/docs/writing-stories/args
-export const Default: Story = {
-  args: {
-    variant: 'default',
     children: (
       <div>
         <CardTitle>Title</CardTitle>
@@ -49,30 +35,10 @@ export const Default: Story = {
   },
 };
 
-export const Clickable: Story = {
-  args: {
-    variant: 'clickable',
-    children: <button className='text-left'>
-      <CardTitle>Title</CardTitle>
-      <CardContent>Content</CardContent>
-      <CardFooter>
-        <Button asChild>
-          <a href="#">Link</a>
-        </Button>
-      </CardFooter>
-    </button>
-  },
-};
+export default meta;
+type Story = StoryObj<typeof meta>;
 
-export const Console: Story = {
-  args: {
-    variant: 'console',
-    children: <div><CardTitle>Title</CardTitle>
-      <CardContent>Content</CardContent>
-      <CardFooter>
-        <ConsoleButton asChild>
-          <a href="#">Link</a>
-        </ConsoleButton>
-      </CardFooter></div>
-  },
+// More on writing stories with args: https://storybook.js.org/docs/writing-stories/args
+export const Default: Story = {
+  args: {},
 };
